Introduced a foreach() plugin.
This is the VQL equivalent of an SQL join, except it is more
intuitive:

SELECT * FROM foreach(
  row={SELECT A, B, C from plugin() },
  query={SELECT x,y,z, A, B, C from repeating_query(arg=A)})

This execute query in "row" and for each row, creates a new scope,
then passes it to the child query. The child query can mix its plugin
results with the child scope - so it is like the row from the "row"
query is presented as args to the child query.