Fixed Default member in OrderedDict bug.
OrderedDicts can now specify a default member. Previously the default
was set when the dict did not contain the key with a pres = true
indication. But this breaks scope walks because the top level object
on the scope will resolve all items and therefore mask the outer
scopes. This change allows an Associative protocol to return a default
value but still indicate that the item is not present. This way the
resolve walk can continue to find a better match and fall back to the
default if there was no better match.