Refactored Reduce() to use less channels and improved group by.
Previously group by was not efficient - it accumulated all binned
rows in memory which is not scalable for large tables.

This change introduces a context within the scope where aggregate
functions may store information from row to row. The context is
created for each group by binned value and so it is possible to
aggregate across the same value (count, min, max etc).

Additionally this change refactors VQL functions to just return their
result directly instead of through a channel. This reduces overall
channels and makes things faster.