Implemented LazyRow for transformed rows.
When evaluating a VQL statement, the FROM clause is generated to
return a row, then this row is transformed by the column
specification, the this transformed flow is filtered by the Where
clause.

Previously the transformed row would evaluate the column expressions
into a Dict before passing them into the Where clause. This means that
expressions with side effects would be run - even if they are later
filtered out. So an expression like:

SELECT upload(file=Exe) from pslist() WHERE Name =~ "cmd.exe"

Previously the upload function would be evaluated before passing to
the WHERE clause which would cuase all processes to be uploaded.

With the new lazy row, we only evaluate the columns when they are
referenced. Hence in the above query, the upload function will only be
matrerialized for those rows which are satisfied by the WHERE clause.