Fixed parser bug with Index
Previously Associative following an Index was not evaluated
correctly. Bug was of the form:

(Foo[0]).Bar

This is required to index the result of the LET expression:

LET Foo <= SELECT ....

Foo will be an array so:

Foo[0] is the first row and (Foo[0]).Bar is the "Bar" column.

Note that it is currently required to use parenthesis here.