Removed the Column() API
The Column API was originally thought to allow plugins to pre-declare
their columns in advance. This was thought to provide documentation
and maybe some optimizations.

It turned out to be not very useful because most plugins do not
declare it and it leads to bugs because it is hard to predict the
columns the plugin actually returns when different from what it
declares.

Removing this API cleans up the code and removes these bugs. Now we
always determine the columns returned by the keys of the first
row. Plugins should always return the same row object every time.