Expand the columns of each row separately.
In theory a result set should have all the same columns for each row
but this seems to be hard to satisfy in practice. For example when
calling an artifact with multiple sources each should may have
different queries.

This change removes thie restriction and expands every row's columns
separately. In the common case where rows are just dicts there should
be no additional performance impact.