Added plugin versioning.
The version() function can retrieve the version of any plugin or
function. This helps to manage plugin versions with VQL being able to
do different code paths depending on plugin version.

If the plugin does not exist at all version() returns NULL - therefore
it is possible to check for existance of plugins too.

example:

```
SELECT * FROM if(
  condition=version(plugin="foo") > 1,
  then=ModernQuery,
  else=LegacyQuery)
```