Added a worker pool to foreach().
A worker pool limits the number of concurrent workers for async
foreach() making it safe to run over a lot of results. By limiting the
number of workers we can match the workers to the number of cpus and
therefore get parallelism for cpu intensive tasks. For example:

```sql
SELECT * FROM foreach(row={SELECT FullPath FROM glob(...)},
   workers=30,
   query={
      SELECT hash(file=FullPath), FullPath FROM scope()
   })
```

This will calculate hashes on 30 cores leading to a huge speedup.