Allow star to come at any place in the column expression. (#84)
Previously * had to come first, optionally followed by other
expressions:

SELECT *, "Extra" AS X FROM plugin()

This caused additional columns to be added to the row or potentially
override existing columns.

This PR makes the following syntax possible

SELECT "Extra" AS X, * FROM plugin()

This allows extra columns to be added to the front to the column
list. The extra columns can still override previous columns but if
they are not they will be placed first.