Automatically allow struct access by json tags (#108)
This fixes issues where non-dicts structs are injected into VQL
output. If a custom struct has both fields and json tags which are not
consistent, then the user sees the json tags as the struct is
serialized. But then when they attempt to access the field by these
this will fail because the struct name is actually different:

type Foo struct {
  SrcIP string `json:"src_ip"`
}

Accessing the field like `X.src_ip` was failing because the struct
does not really have a field `src_ip` - even though it looks like it
because the json serialized data does!